Finished test mode with skeleton final code.

+++ lighting_patterns/Final Code/test/Graph.hpp
@@ -21,7 +21,7 @@ public:
      * technicaly edges is a nV x something matrix
      * TODO change edges to an array of pointers
      */
-    Graph(int edges[291][291 + 1], int nV) : _matrix() {
+    Graph(int16_t edges[291][291 + 1], int nV) : _matrix() {
         vector<bool> *tmp = new vector<bool> (nV, false);
         _matrix.resize(nV, *tmp);
         delete tmp;
